+++ test/pagy/extras/arel_test.rb
@@ -4,16 +4,13 @@ require_relative '../../test_helper'
 require_relative '../../mock_helpers/arel'
 require 'pagy/extras/arel'
 
-describe Pagy::Backend do
-
+describe 'pagy/extras/arel' do
   let(:controller) { MockController.new }
 
   describe '#pagy_arel' do
-
     before do
       @collection = MockCollection.new
     end
-
     it 'paginates with defaults' do
       pagy, items = controller.send(:pagy_arel, @collection)
       _(pagy).must_be_instance_of Pagy
@@ -23,7 +20,6 @@ describe Pagy::Backend do
       _(items.size).must_equal Pagy::VARS[:items]
       _(items).must_equal [41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60]
     end
-
     it 'paginates with vars' do
       pagy, items = controller.send(:pagy_arel, @collection, page: 2, items: 10, link_extra: 'X')
       _(pagy).must_be_instance_of Pagy
@@ -34,15 +30,12 @@ describe Pagy::Backend do
       _(items.size).must_equal 10
       _(items).must_equal [11, 12, 13, 14, 15, 16, 17, 18, 19, 20]
     end
-
   end
 
   describe '#pagy_arel_get_vars' do
-
     before do
       @collection = MockCollection.new
     end
-
     it 'gets defaults' do
       vars   = {}
       merged = controller.send :pagy_arel_get_vars, @collection, vars
@@ -51,7 +44,6 @@ describe Pagy::Backend do
       _(merged[:count]).must_equal 1000
       _(merged[:page]).must_equal 3
     end
-
     it 'gets vars' do
       vars   = {page: 2, items: 10, link_extra: 'X'}
       merged = controller.send :pagy_arel_get_vars, @collection, vars
@@ -64,7 +56,6 @@ describe Pagy::Backend do
       _(merged[:items]).must_equal 10
       _(merged[:link_extra]).must_equal 'X'
     end
-
     it 'works with grouped collections' do
       @collection = MockCollection::Grouped.new((1..1000).to_a)
       vars   = {page: 2, items: 10, link_extra: 'X'}
@@ -78,7 +69,6 @@ describe Pagy::Backend do
       _(merged[:items]).must_equal 10
       _(merged[:link_extra]).must_equal 'X'
     end
-
     it 'overrides count and page' do
       vars   = {count: 10, page: 32}
       merged = controller.send :pagy_arel_get_vars, @collection, vars
@@ -87,7 +77,5 @@ describe Pagy::Backend do
       _(merged.keys).must_include :page
       _(merged[:page]).must_equal 32
     end
-
   end
-
 end