+++ test/pagy/extras/oj_shared_test.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'oj'
-require_relative '../../test_helper'
-require 'pagy/extras/shared'
-
-require_relative 'shared_test'