+++ test/pagy_test.rb
@@ -2,20 +2,35 @@
 
 require_relative 'test_helper'
 
-describe Pagy do
+def series_for(page, *expected)
+  expected.each_with_index do |value, index|
+    vars = instance_variable_get(:"@vars#{index}").merge(page: page)
+    _(Pagy.new(vars).series).must_equal value
+  end
+end
 
+describe 'pagy' do
   let(:pagy) { Pagy.new count: 100, page: 4 }
 
-  it 'has version' do
-    _(Pagy::VERSION).wont_be_nil
+  describe 'version match' do
+    it 'has version' do
+      _(Pagy::VERSION).wont_be_nil
+    end
+    it 'defines the same version in config/pagy.rb' do
+      _(File.read(Pagy.root.join('config', 'pagy.rb'))).must_match "# Pagy initializer file (#{Pagy::VERSION})"
+    end
+    it 'defines the same version in javascripts/pagy.js' do
+      _(File.read(Pagy.root.join('javascripts', 'pagy.js'))).must_match "Pagy.version = '#{Pagy::VERSION}';"
+    end
+    it 'defines the same version in CHANGELOG.md' do
+      _(File.read(Pagy.root.parent.join('CHANGELOG.md'))).must_match "## Version #{Pagy::VERSION}"
+    end
   end
 
   describe '#initialize' do
-
     before do
       @vars  = { count: 103, items: 10, size: [3, 2, 2, 3] }
     end
-
     it 'initializes' do
       _(pagy).must_be_instance_of Pagy
       _(Pagy.new(count: 100)).must_be_instance_of Pagy
@@ -37,7 +52,6 @@ describe Pagy do
         _(e.pagy).must_be_instance_of Pagy
       end
     end
-
     it 'initializes count 0' do
       pagy = Pagy.new @vars.merge(count: 0)
       _(pagy.pages).must_equal 1
@@ -48,7 +62,6 @@ describe Pagy do
       _(pagy.prev).must_be_nil
       _(pagy.next).must_be_nil
     end
-
     it 'initializes single page' do
       pagy = Pagy.new @vars.merge(count: 8)
       _(pagy.pages).must_equal 1
@@ -59,7 +72,6 @@ describe Pagy do
       _(pagy.prev).must_be_nil
       _(pagy.next).must_be_nil
     end
-
     it 'initializes page 1 of 2' do
       pagy = Pagy.new @vars.merge(count: 15)
       _(pagy.pages).must_equal 2
@@ -70,7 +82,6 @@ describe Pagy do
       _(pagy.prev).must_be_nil
       _(pagy.next).must_equal 2
     end
-
     it 'initializes page 2 of 2' do
       pagy = Pagy.new @vars.merge(count: 15, page: 2)
       _(pagy.pages).must_equal 2
@@ -82,7 +93,6 @@ describe Pagy do
       _(pagy.page).must_equal 2
       _(pagy.next).must_be_nil
     end
-
     it 'initializes page 1' do
       pagy = Pagy.new @vars.merge(page: 1)
       _(pagy.count).must_equal 103
@@ -96,7 +106,6 @@ describe Pagy do
       _(pagy.page).must_equal 1
       _(pagy.next).must_equal 2
     end
-
     it 'initializes page 2' do
       pagy = Pagy.new @vars.merge(page: 2)
       _(pagy.count).must_equal 103
@@ -110,7 +119,6 @@ describe Pagy do
       _(pagy.page).must_equal 2
       _(pagy.next).must_equal 3
     end
-
     it 'initializes page 3' do
       pagy = Pagy.new @vars.merge(page: 3)
       _(pagy.count).must_equal 103
@@ -124,7 +132,6 @@ describe Pagy do
       _(pagy.page).must_equal 3
       _(pagy.next).must_equal 4
     end
-
     it 'initializes page 4' do
       pagy = Pagy.new @vars.merge(page: 4)
       _(pagy.count).must_equal 103
@@ -138,7 +145,6 @@ describe Pagy do
       _(pagy.page).must_equal 4
       _(pagy.next).must_equal 5
     end
-
     it 'initializes page 5' do
       pagy = Pagy.new @vars.merge(page: 5)
       _(pagy.count).must_equal 103
@@ -152,7 +158,6 @@ describe Pagy do
       _(pagy.page).must_equal 5
       _(pagy.next).must_equal 6
     end
-
     it 'initializes page 6' do
       pagy = Pagy.new @vars.merge(page: 6)
       _(pagy.count).must_equal 103
@@ -166,7 +171,6 @@ describe Pagy do
       _(pagy.page).must_equal 6
       _(pagy.next).must_equal 7
     end
-
     it 'initializes page 7' do
       pagy = Pagy.new @vars.merge(page: 7)
       _(pagy.count).must_equal 103
@@ -180,7 +184,6 @@ describe Pagy do
       _(pagy.page).must_equal 7
       _(pagy.next).must_equal 8
     end
-
     it 'initializes page 8' do
       pagy = Pagy.new @vars.merge(page: 8)
       _(pagy.count).must_equal 103
@@ -194,7 +197,6 @@ describe Pagy do
       _(pagy.page).must_equal 8
       _(pagy.next).must_equal 9
     end
-
     it 'initializes page 9' do
       pagy = Pagy.new @vars.merge(page: 9)
       _(pagy.count).must_equal 103
@@ -208,7 +210,6 @@ describe Pagy do
       _(pagy.page).must_equal 9
       _(pagy.next).must_equal 10
     end
-
     it 'initializes page 10' do
       pagy = Pagy.new @vars.merge(page: 10)
       _(pagy.count).must_equal 103
@@ -222,7 +223,6 @@ describe Pagy do
       _(pagy.page).must_equal 10
       _(pagy.next).must_equal 11
     end
-
     it 'initializes page 11' do
       pagy = Pagy.new @vars.merge(page: 11)
       _(pagy.count).must_equal 103
@@ -236,7 +236,6 @@ describe Pagy do
       _(pagy.page).must_equal 11
       _(pagy.next).must_be_nil
     end
-
     it 'initializes outset page 1' do
       pagy = Pagy.new(count: 87, page: 1, outset: 10, items: 10)
       _(pagy.offset).must_equal 10
@@ -245,7 +244,6 @@ describe Pagy do
       _(pagy.to).must_equal 10
       _(pagy.pages).must_equal 9
     end
-
     it 'initializes outset page 9' do
       pagy = Pagy.new(count: 87, page: 9, outset: 10, items: 10)
       _(pagy.offset).must_equal 90
@@ -254,28 +252,23 @@ describe Pagy do
       _(pagy.to).must_equal 87
       _(pagy.pages).must_equal 9
     end
-
     it 'initializes items of last page of one' do
       _(Pagy.new(items: 10, count: 0).items).must_equal 10
       _(Pagy.new(items: 10, count: 4).items).must_equal 4
       _(Pagy.new(items: 10, count: 10).items).must_equal 10
     end
-
     it 'initializes items of last page of many' do
       _(Pagy.new(items: 10, count: 14, page: 2).items).must_equal 4
       _(Pagy.new(items: 10, count: 20, page: 2).items).must_equal 10
     end
-
     it 'handles the :cycle variable' do
       pagy = Pagy.new(count: 100, page: 10, items: 10, cycle: true)
       _(pagy.prev).must_equal(9)
       _(pagy.next).must_equal 1
     end
-
   end
 
   describe 'accessors' do
-
     it 'has accessors' do
       [
       :count, :page, :items, :vars, # input
@@ -284,11 +277,9 @@ describe Pagy do
         _(pagy).must_respond_to meth
       end
     end
-
   end
 
   describe 'variables' do
-
     it 'has vars defaults' do
       _(Pagy::VARS[:page]).must_equal 1
       _(Pagy::VARS[:items]).must_equal 20
@@ -297,11 +288,9 @@ describe Pagy do
       _(Pagy::VARS[:page_param]).must_equal :page
       _(Pagy::VARS[:params]).must_equal({})
     end
-
   end
 
   describe '#series' do
-
     before do
       @vars0 = { count: 103,
                  items: 10,
@@ -316,14 +305,6 @@ describe Pagy do
                  items: 10,
                  size:  [3, 2, 2, 3] }
     end
-
-    def series_for(page, *expected)
-      expected.each_with_index do |value, index|
-        vars = instance_variable_get(:"@vars#{index}").merge(page: page)
-        _(Pagy.new(vars).series).must_equal value
-      end
-    end
-
     it 'computes series for page 1' do
       series_for 1,
                  ["1", 2, 3, :gap],
@@ -331,7 +312,6 @@ describe Pagy do
                  ["1", 2, 3, :gap],
                  ["1", 2, 3, :gap, 9, 10, 11]
     end
-
     it 'computes series for page 2' do
       series_for 2,
                  [1, "2", 3, 4, :gap],
@@ -339,7 +319,6 @@ describe Pagy do
                  [1, "2", 3, :gap],
                  [1, "2", 3, 4, :gap, 9, 10, 11]
     end
-
     it 'computes series for page 3' do
       series_for 3,
                  [1, 2, "3", 4, 5, :gap],
@@ -347,7 +326,6 @@ describe Pagy do
                  [1, 2, "3", :gap],
                  [1, 2, "3", 4, 5, :gap, 9, 10, 11]
     end
-
     it 'computes series for page 4' do
       series_for 4,
                  [1, 2, 3, "4", 5, 6, :gap],
@@ -355,7 +333,6 @@ describe Pagy do
                  [1, 2, 3, "4", :gap],
                  [1, 2, 3, "4", 5, 6, :gap, 9, 10, 11]
     end
-
     it 'computes series for page 5' do
       series_for 5,
                  [:gap, 3, 4, "5", 6, 7, :gap],
@@ -363,7 +340,6 @@ describe Pagy do
                  [1, 2, 3, 4, "5", :gap],
                  [1, 2, 3, 4, "5", 6, 7, 8, 9, 10, 11]
     end
-
     it 'computes series for page 6' do
       series_for 6,
                  [:gap, 4, 5, "6", 7, 8, :gap],
@@ -371,7 +347,6 @@ describe Pagy do
                  [1, 2, 3, 4, 5, "6", :gap],
                  [1, 2, 3, 4, 5, "6", 7, 8, 9, 10, 11]
     end
-
     it 'computes series for page 7' do
       series_for 7,
                  [:gap, 5, 6, "7", 8, 9, :gap],
@@ -379,7 +354,6 @@ describe Pagy do
                  [1, 2, 3, 4, 5, 6, "7", :gap],
                  [1, 2, 3, 4, 5, 6, "7", 8, 9, 10, 11]
     end
-
     it 'computes series for page 8' do
       series_for 8,
                  [:gap, 6, 7, "8", 9, 10, 11],
@@ -387,7 +361,6 @@ describe Pagy do
                  [1, 2, 3, :gap, 6, 7, "8", :gap],
                  [1, 2, 3, :gap, 6, 7, "8", 9, 10, 11]
     end
-
     it 'computes series for page 9' do
       series_for 9,
                  [:gap, 7, 8, "9", 10, 11],
@@ -395,7 +368,6 @@ describe Pagy do
                  [1, 2, 3, :gap, 7, 8, "9", :gap],
                  [1, 2, 3, :gap, 7, 8, "9", 10, 11]
     end
-
     it 'computes series for page 10' do
       series_for 10,
                  [:gap, 8, 9, "10", 11],
@@ -403,7 +375,6 @@ describe Pagy do
                  [1, 2, 3, :gap, 8, 9, "10", 11],
                  [1, 2, 3, :gap, 8, 9, "10", 11]
     end
-
     it 'computes series for page 11' do
       series_for 11,
                  [:gap, 9, 10, "11"],
@@ -411,27 +382,20 @@ describe Pagy do
                  [1, 2, 3, :gap, 9, 10, "11"],
                  [1, 2, 3, :gap, 9, 10, "11"]
     end
-
     it 'computes series for count 0' do
       _(Pagy.new(@vars3.merge(count: 0)).series).must_equal ["1"]
     end
-
     it 'computes series for single page' do
       _(Pagy.new(@vars3.merge(count: 8)).series).must_equal ["1"]
     end
-
     it 'computes series for 1 of 2 pages' do
       _(Pagy.new(@vars3.merge(count: 15)).series).must_equal ["1", 2]
     end
-
     it 'computes series for 2 of 2 pages' do
       _(Pagy.new(@vars3.merge(count: 15, page: 2)).series).must_equal [1, "2"]
     end
-
     it 'computes an empty series' do
       _(Pagy.new(@vars3.merge(count: 100, size: [])).series).must_equal []
     end
-
   end
-
 end